#1: Refactor namespaces for better organization and clarity in GameInfo, Plugin, and PluginInfo classes
