+++ src/GameInfo.cs
@@ -1,9 +1,9 @@
-namespace RepoEssentials;
-
-public static class GameInfo {
-    public const string NAME = "R.E.P.O.";
-    public const string VERSION = "0.1.2";
-    public const string APP_ID = "3241660";
-    public const string BUILD_ID = "17560228";
-    public const string EXECUTABLE_NAME = "REPO.exe";
+namespace RepoEssentials.src {
+    public static class GameInfo {
+        public const string NAME = "R.E.P.O.";
+        public const string VERSION = "0.1.2";
+        public const string APP_ID = "3241660";
+        public const string BUILD_ID = "17560228";
+        public const string EXECUTABLE_NAME = "REPO.exe";
+    }
 }
