+++ src/Plugin.cs
@@ -2,26 +2,25 @@
 using BepInEx.Logging;
 using HarmonyLib;
 
-namespace RepoEssentials;
+namespace RepoEssentials.src {
+    [BepInPlugin(PluginInfo.GUID, PluginInfo.NAME, PluginInfo.VERSION)]
+    [BepInProcess(GameInfo.EXECUTABLE_NAME)]
+    public class Plugin : BaseUnityPlugin {
+        internal static new ManualLogSource Logger;
 
+        private void Awake() {
+            // Plugin startup logic
+            Logger = base.Logger;
+            Logger.LogInfo("Loading plugin...");
 
-[BepInPlugin(PluginInfo.GUID, PluginInfo.NAME, PluginInfo.VERSION)]
-[BepInProcess(GameInfo.EXECUTABLE_NAME)]
-public class Plugin : BaseUnityPlugin {
-    internal static new ManualLogSource Logger;
+            // Configure Harmony
+            Harmony harmony = new(PluginInfo.GUID);
 
-    private void Awake() {
-        // Plugin startup logic
-        Logger = base.Logger;
-        Logger.LogInfo("Loading plugin...");
+            // Apply patches
+            harmony.PatchAll();
+            Logger.LogInfo("Harmony patches are loaded!");
 
-        // Configure Harmony
-        Harmony harmony = new(PluginInfo.GUID);
-
-        // Apply patches
-        harmony.PatchAll();
-        Logger.LogInfo("Harmony patches are loaded!");
-
-        Logger.LogInfo("Plugin loaded successfully!");
+            Logger.LogInfo("Plugin loaded successfully!");
+        }
     }
 }
