+++ src/PluginInfo.cs
@@ -1,11 +1,11 @@
-namespace RepoEssentials;
-
-public static class PluginInfo {
-    public const string GUID = "org.ccarrmcmahon.plugins.repo.essentials";
-    public const string NAME = "Essentials";
-    public const string VERSION = "0.1.0";
-    public const string AUTHOR = "CCarrMcMahon";
-    public const string ENTRY = "RepoEssentials.Plugin.Awake";
-    public const string MANIFEST = "manifest.json";
-    public const string ASSEMBLY = "Essentials.dll";
+namespace RepoEssentials.src {
+    public static class PluginInfo {
+        public const string GUID = "org.ccarrmcmahon.plugins.repo.essentials";
+        public const string NAME = "Essentials";
+        public const string VERSION = "0.1.0";
+        public const string AUTHOR = "CCarrMcMahon";
+        public const string ENTRY = "RepoEssentials.Plugin.Awake";
+        public const string MANIFEST = "manifest.json";
+        public const string ASSEMBLY = "Essentials.dll";
+    }
 }
